automaizo la eliminacion de  twitcher data

+++ uninstall.php
@@ -3,10 +3,17 @@
 if(!defined('WP_UNINSTALL_PLUGIN')){
     die();
 }
-// if (twchr delte all == true)
-delete_option('twchr_keys');
-delete_option('twchr_app_token');
-delete_option('twchr_setInstaled');
-delete_option('twchr_data_broadcaster');
+if (get_option('twchr_delete_all') == 1){
+    delete_option('twchr_keys');
+    delete_option('twchr_app_token');
+    delete_option('twchr_setInstaled');
+    delete_option('twchr_data_broadcaster');
+
+    $allposts= get_posts( array('post_type'=>'twchr_streams','numberposts'=>-1) );
+    foreach ($allposts as $eachpost) {
+        wp_delete_post( $eachpost->ID, true );
+    }
+}
+
 
 ?>
